Configure unicorn

+++ config/unicorn.rb
@@ -1,59 +1,42 @@
-# number of workers
+# Set your full path to application.
+app_dir = File.expand_path('../../', __FILE__)
+shared_dir = File.expand_path('../../../shared/', __FILE__)
+
+# Set unicorn options
 worker_processes 2
-
-# listen on a unix socket
-listen "#{Dir.pwd}/tmp/sockets/unicorn.sock"
-
-# define a pid for the master process
-pid "#{Dir.pwd}/tmp/pids/unicorn.pid"
-
-# By default, the Unicorn logger will write to stderr.
-# Additionally, some applications/frameworks log to stderr or stdout,
-# so prevent them from going to /dev/null when daemonized here:
-stderr_path "#{Dir.pwd}/log/unicorn.stderr.log"
-stdout_path "#{Dir.pwd}/log/unicorn.stdout.log"
-
-# preload_app true is required for New Relic and REE
 preload_app true
+timeout 30
+
+# Fill path to your app
+working_directory app_dir
+
+# Set up socket location
+listen "#{shared_dir}/sockets/unicorn.sock", backlog: 64
+
+# Logging
+stderr_path "#{shared_dir}/log/unicorn.stderr.log"
+stdout_path "#{shared_dir}/log/unicorn.stdout.log"
+
+# Set master PID location
+pid "#{shared_dir}/pids/unicorn.pid"
 
 before_fork do |server, worker|
-  # the following is highly recomended for Rails + "preload_app true"
-  # as there's no need for the master process to hold a connection
-  defined?(ActiveRecord::Base) and
-    ActiveRecord::Base.connection.disconnect!
-
-  # The following is only recommended for memory/DB-constrained
-  # installations. It is not needed if your system can house
-  # twice as many worker_processes as you have configured.
-  #
-  # This allows a new master process to incrementally
-  # phase out the old master process with SIGTTOU to avoid a
-  # thundering herd (especially in the "preload_app false" case)
-  # when doing a transparent upgrade.  The last worker spawned
-  # will then kill off the old master process with a SIGQUIT.
-  #
-  # executed via 'service myinit reload'
+  defined?(ActiveRecord::Base) and ActiveRecord::Base.connection.disconnect!
   old_pid = "#{server.config[:pid]}.oldbin"
-  if old_pid != server.pid
+  if File.exists?(old_pid) && server.pid != old_pid
     begin
       sig = (worker.nr + 1) >= server.worker_processes ? :QUIT : :TTOU
       Process.kill(sig, File.read(old_pid).to_i)
     rescue Errno::ENOENT, Errno::ESRCH
+      # someone else did our job for us
     end
   end
-  #
-  # *optionally* throttle the master from forking too quickly by sleeping
-  # sleep 1
 end
 
 after_fork do |server, worker|
-  # the following is *required* for Rails + "preload_app true",
-  defined?(ActiveRecord::Base) and
-    ActiveRecord::Base.establish_connection
+  defined?(ActiveRecord::Base) and ActiveRecord::Base.establish_connection
+end
 
-  # if preload_app is true, then you may also want to check and
-  # restart any other shared sockets/descriptors such as Memcached,
-  # and Redis.  TokyoCabinet file handles are safe to reuse
-  # between any number of forked children (assuming your kernel
-  # correctly implements pread()/pwrite() system calls)
+before_exec do |server|
+  ENV['BUNDLE_GEMFILE'] = "#{app_dir}/Gemfile"
 end
